requirements.txt
pytest==6.2.5
pytest-selenium==4.0.0
virtualenv
Appium-Python-Client~=2.7.0
AltTester-Driver==1.8.0
assertpy
pytest-html
requests~=2.28.1
PyYAML~=6.0
dominate~=2.7.0